pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "lacrei-saude-api"
version = "0.1.0"
description = "API RESTful de Gerenciamento de Consultas Médicas"
authors = [
    {name = "Olavo Licursi",email = "licursiolavo@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-filter (>=25.2,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
